refactor: move files to their respective folders to organize them
