+++ link_analysis/run
@@ -1,10 +1,10 @@
 #!/bin/sh
 
-g++ *.cpp *.h
+g++ ./src/*.cpp ./includes/*.h ./tests/main.cpp
 
 # Run code
 ./a.out
 
-# Clean up
+# Clean up the binary and the .gch files created from header files
 rm a.out
-rm *.gch
+rm ./includes/*.gch
