--- link_analysis/tests/test.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/bash
-
-g++ ../graph.cpp ../graph.h main.cpp
-./a.out
